+++ Unit 04/Pathway 02/basic-android-kotlin-compose-training-cupcake/app/src/main/java/com/example/cupcake/CupcakeScreen.kt
@@ -75,8 +75,11 @@ enum class CupcakeScreen(@StringRes val title: Int) {
 }
 
 @Composable
-fun CupcakeApp(modifier: Modifier = Modifier, viewModel: OrderViewModel = viewModel()) {
-    val navController = rememberNavController()
+fun CupcakeApp(
+    modifier: Modifier = Modifier,
+    viewModel: OrderViewModel = viewModel(),
+    navController: NavHostController = rememberNavController()
+) {
 
     val backStackEntry by navController.currentBackStackEntryAsState()